Move cache migration to separate module to simplify code review.
